error-kb-agent/pyproject.toml
[project]
name = "error-kb-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk[a2a]>=1.28.1",
    "python-dotenv>=1.2.2",
    "toolbox-adk>=1.0.0",
    "toolbox-core>=1.0.0",
]
